feat(model-e3): OT feature update evaluation logic change
Make OT feature +1 study with a small probabiltiy and +1 test and +1 between lists

Problem found and this did solved the problem

The winning list problem was caused by...during the real time of the testing the sampled probe will be marked as tested. And if an error happened earlier in mistakenly sampled the current list later targets as a false alarm, let's say, then that target will be marked as tested. And according to the decision rule, when it's marked as tested, it will be judged as new rather than old. That means earlier tests of the current lists potentially makes later target tests of the current list to go worse. And this is the reverse of the observation.

Closes #57